Don't use HTML5 Renderer for tikz test
We instead use a custom renderer designed for this test. This means the
test will not fail if we decide to change how HTML5 displays
tikz/tikzcd.

Currently, the test ensures that when we access `obj.vectorImage.url`
and `obj.image.url`, the images are produced correctly and the correct
url is returned.
