+++ unittests/benchmarks/tikz-pdf2svg-gspdfpng/index
@@ -1,0 +1,4 @@
+images/img-0001.svg
+images/img-0001.png
+images/img-0002.svg
+images/img-0002.png
