--- unittests/benchmarks/tikz-pdf2svg-gspdfpng/index.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-<script type="text/x-mathjax-config">
-  MathJax.Hub.Config({tex2jax: {inlineMath: [ ['\\(','\\)']]}});
-</script>
-<script type="text/javascript" src="http://cdn.mathjax.org/mathjax/latest/MathJax.js?config=TeX-AMS_CHTML">
-</script>
-<meta name="generator" content="plasTeX" />
-<meta charset="utf-8" />
-<meta name="viewport" content="width=device-width, initial-scale=1" />
-<link rel="stylesheet" href="styles/theme-green.css" />
-</head>
-
-<body>
-
-<div class="wrapper">
-
-
-<div class="content">
-<div class="tikzpicture" id="a0000000002">
-    <img src="images/img-0001.svg" alt="\begin{tikzpicture} 
-        \draw (0, 0) -- (0, 2) -- (2, 0) -- (0, 0);
-    \end{tikzpicture}" />
-</div>
-<div class="tikzcd" id="a0000000003">
-    <img src="images/img-0002.svg" alt="\begin{tikzcd} 
-        A \ar[r, &#34;i&#34;] &amp; B \ar[d, &#34;p&#34;] \\ &amp; C
-    \end{tikzcd}" />
-</div>
-
-</div>
-
-<nav class="prev_up_next">
-</nav>
-
-<script type="text/javascript" src="js/jquery.min.js"></script>
-<script type="text/javascript" src="js/plastex.js"></script>
-<script type="text/javascript" src="js/svgxuse.js"></script>
-</div>
-</body>
-</html>
